Reimplement PBKDF2SHA512 using Java 8 Crypto package
The original PBKDF2SHA512 is a cryptofreek.org implementation of pbkdf2.
The java doc is copied from the cryptofreek.org post, and states that
the class passes all the RFC 6070 test vectors. This is not technically
true for PBKDF2SHA512 because we changed the SecretKeyFactory algorithm
from HmacSHA1 to HmacSHA512.

The Java 8 implementation of "PBKDF2WithHmacSHA512" is used to satisfy
the original method signature for derive(...).